Exit lightbox overlay mode before animation.
The exit animation is now positioned with `position: absolute`, so you can
scroll during the animation and it will still animate correctly. This
allows the user to scroll immediately after closing the lightbox.
Previously, there was a small, but noticeable delay where you could try to scroll
and nothing would happen.